% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philart.R
\name{philart_artworks}
\alias{philart_artworks}
\title{Browse multiple artworks with detail}
\usage{
philart_artworks(limit = 20, offset = 0)
}
\arguments{
\item{limit}{Number of artworks to fetch (default 20)}
\item{offset}{Number of artworks to skip (default 0)}
}
\value{
tibble: id, title, artists, year, latitude, longitude,
  location_description, architecture, description, image_url
}
\description{
Browse multiple artworks with detail
}
