% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philart.R
\name{philart_detail}
\alias{philart_detail}
\title{Get detail for a single artwork by ID}
\usage{
philart_detail(id)
}
\arguments{
\item{id}{Artwork ID (integer)}
}
\value{
tibble: single row with id, title, artists, year, latitude,
  longitude, location_description, architecture, description, image_url
}
\description{
Get detail for a single artwork by ID
}
