% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philart.R
\name{philart_search}
\alias{philart_search}
\title{Search artworks by keyword}
\usage{
philart_search(query)
}
\arguments{
\item{query}{Search string (case-insensitive)}
}
\value{
tibble: id, name
}
\description{
Search artworks by keyword
}
